Move template specific utilities into module
