+++ qubesadmin/templates.py
@@ -20,6 +20,7 @@
 """Template management for Qubes OS."""
 
 import os
+import fnmatch
 import subprocess
 import tempfile
 import typing
@@ -30,6 +31,73 @@ import qubesadmin.exc
 PATH_PREFIX = '/var/lib/qubes/vm-templates'
 PACKAGE_NAME_PREFIX = 'qubes-template-'
 TAR_HEADER_BYTES = 512
+
+
+def qubes_release() -> str:
+    """Return the Qubes release."""
+    if os.path.exists('/usr/share/qubes/marker-vm'):
+        with open('/usr/share/qubes/marker-vm', 'r', encoding='ascii') as fd:
+            # Get the first non-comment line
+            release = [l.strip() for l in fd.readlines()
+                       if l.strip() and not l.startswith('#')]
+            # sanity check
+            if release and release[0] and release[0][0].isdigit():
+                return release[0]
+    with open('/etc/os-release', 'r', encoding='ascii') as fd:
+        release = None
+        distro_id = None
+        for line in fd:
+            line = line.strip()
+            if not line or line[0] == '#':
+                continue
+            key, val = line.split('=', 1)
+            if key == 'ID':
+                distro_id = val
+            if key == 'VERSION_ID':
+                release = val.strip('\'"') # strip possible quotes
+        if distro_id and 'qubes' in distro_id and release:
+            return release
+    # Return default value instead of throwing so that it works on CI
+    return '4.1'
+
+
+def build_version_str(evr: typing.Tuple[str, str, str]) -> str:
+    """Return version string described by ``evr``, which is in (epoch, version,
+    release) format."""
+    return '%s:%s-%s' % evr
+
+
+def is_match_spec(name: str, epoch: str, version: str, release: str, spec: str
+                  ) -> typing.Tuple[bool, float]:
+    """Check whether (name, epoch, version, release) matches the spec string.
+
+    For the algorithm, refer to section "NEVRA Matching" in the DNF
+    documentation.
+
+    Note that currently ``arch`` is ignored as the templates should be of
+    ``noarch``.
+
+    :return: A tuple. The first element indicates whether there is a match; the
+        second element represents the priority of the match (lower is better)
+    """
+    if epoch != '0':
+        targets = [
+            f'{name}-{epoch}:{version}-{release}',
+            f'{name}',
+            f'{name}-{epoch}:{version}'
+        ]
+    else:
+        targets = [
+            f'{name}-{epoch}:{version}-{release}',
+            f'{name}-{version}-{release}',
+            f'{name}',
+            f'{name}-{epoch}:{version}',
+            f'{name}-{version}'
+        ]
+    for prio, target in enumerate(targets):
+        if fnmatch.fnmatch(target, spec):
+            return True, prio
+    return False, float('inf')
 
 
 def verify_rpm(path: str, key: str, *, nogpgcheck: bool = False,
